+++ src/components/work/player.js
@@ -4,14 +4,22 @@ import { StyleSheet, ScrollView } from 'react-native';
 const styles = StyleSheet.create({
   container: {
     flex: 1,
+    backgroundColor: 'white',
+  },
+  contentContainer: {
+    flex: 0,
     justifyContent: 'flex-start',
     alignItems: 'center',
-    backgroundColor: 'white',
+    paddingTop: 25,
+    paddingHorizontal: 20,
   },
 });
 
 const WorkPlayer = ({ children }) => (
-  <ScrollView style={{ flex: 1 }} contentContainerStyle={styles.container}>
+  <ScrollView
+    style={styles.container}
+    contentContainerStyle={styles.contentContainer}
+  >
     {children}
   </ScrollView>
 );